Several fixes and improvements
- Added inadmissible as a permanent options instead that on-question
option.
- Help info about inadmissible
- Fixed a bug due to registered processes. Previously, all was in the
same node, so calls to where_is/1 were perfectly working. Now, they are
in a different node, so the where_is/1 should be done in that new node.
We have added a synthtised case expression into the instrumented code
of the send function to do that.
- Improved the way a not selectable event is treated.
- No need to cover anymore send_sent events where the receiver is an
atom.
- Fixed a bug that made unrelated traced events to leave some holes in
the communications diagram.